Moved polygon type selection to polygon panel
