+++ src/Forvaltningsportalen/FeatureInfo/InfoboxSide.js
@@ -56,7 +56,7 @@ const InfoBox = ({
   const [screenHeight, setScreenHeight] = useState(0);
   const [inTransition, setInTransition] = useState(null);
   const [showMarkerOptions, setShowMarkerOptions] = useState(false);
-  // const [showPolygonOptions, setShowPolygonOptions] = useState(false);
+  const [showPolygonOptions, setShowPolygonOptions] = useState(false);
 
   useEffect(() => {
     if (DY < 0 && Y !== 0) {
@@ -277,6 +277,8 @@ const InfoBox = ({
           handlePolygonResults={handlePolygonResults}
           grensePolygon={grensePolygon}
           handleGrensePolygon={handleGrensePolygon}
+          showPolygonOptions={showPolygonOptions}
+          setShowPolygonOptions={setShowPolygonOptions}
         />
       )}
     </div>
